Add a NotFound error.
I feel there's a mismatch at the file::safe_open boundary.  I probably
ought to be in HttpError by that point instead of io::Error.  This is
shifting the conversion closer to that point, but not all the way.  It
improves logging markedly, because (as I was disappointed to note)
io::Error.description() does not return the message text you feed in at
creation.