server: send un-gzipped version's mtime.
In my site's workflow, the gzipped and not-gzipped versions happen to
have equal mtimes, but that's not something we should be relying on.
It's clearest to send the non-gzipped version's mtime for cache control
purposes.

I've also added some explicit logging when the gzipped version is
selected, which is technically redundant.