Revisit error propagation and reporting.
This brings the HttpError type into closer conformance with HTTP's error
model.  We now have a clear association between error values and HTTP
error conditions, and centralizing mapping.

I've also expanded the ability to report log information through errors
without worrying about it being sent to the client.  The logic for this
is unfortunately error-prone, but at least it's all in one place.